pyproject.toml
[project]
name = "kids-storybook-generator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "accelerate>=1.7.0",
    "diffusers>=0.33.1",
    "fpdf>=1.7.2",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "streamlit>=1.45.1",
    "torch>=2.7.1",
    "transformers>=4.52.4",
]
